arch-riscv: Add RV64 vector extension support
This patch adds support for RISC-V V extensions. Currently about 95% of the V extension commands have been implemented in this patch. This implementation uses microops to support all CPU models with minor changes in fetch stages of DerivO3CPU and MinorCPU.

There is a little work to do to fully implement the V extension, including "Vector Compress Instruction", "Fault-Only-First Loads" and "Vector Load/Store Segment Instructions". It is hard to implement these instructions using only microops without microarch modifications of CPU models.

We hope to hear from the community as soon as possible, so we are now open for review.

Change-Id: I7cec10412c00a35c247299cd92d83cdee9066410
Co-authored-by: Xuan Hu <huxuan21@bosc.ac.cn>
Co-authored-by: Yang Liu <numbksco@gmail.com>
Co-authored-by: Fan Yang <1209202421@qq.com>
Co-authored-by: Jerin Joy <joy@rivosinc.com>